+++ tuidb.sql
@@ -130,8 +130,8 @@ CREATE TABLE `location` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `city_id` int(11) NOT NULL,
   `address` text NOT NULL,
-  `x_coord` int(11) NOT NULL,
-  `y_coord` int(11) NOT NULL,
+  `x_coord` double NOT NULL,
+  `y_coord` double NOT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB AUTO_INCREMENT=16 DEFAULT CHARSET=utf8;
 /*!40101 SET character_set_client = @saved_cs_client */;
